fix: add rust_decimal tyeof or decimal sql

+++ quantum_types/src/types/db/superproof.rs
@@ -1,4 +1,5 @@
 use chrono::NaiveDateTime;
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 
 use crate::enums::superproof_status::SuperproofStatus;
@@ -9,11 +10,11 @@ pub struct Superproof {
     pub proof_ids: Option<String>,
     pub superproof_proof_path: Option<String>,
     pub transaction_hash: Option<String>,
-    pub gas_cost: Option<f64>,
+    pub gas_cost: Option<Decimal>,
     pub agg_time: Option<u64>,
     pub status: SuperproofStatus,
     pub superproof_root: Option<String>,
     pub superproof_leaves_path: Option<String>,
     pub onchain_submission_time: Option<NaiveDateTime>,
-    pub eth_price: Option<f64>
+    pub eth_price: Option<Decimal>
 }
